solve the cors error and a board page
